# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses": {
      "post": {
        "operationId": "CourseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "get": {
        "operationId": "CourseController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "operationId": "CourseController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "put": {
        "operationId": "CourseController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "delete": {
        "operationId": "CourseController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
    "/login": {
      "post": {
        "operationId": "AuthenticationController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/select-course": {
      "patch": {
        "operationId": "AuthenticationController_selectCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectCourseReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/ppis": {
      "post": {
        "operationId": "PPIController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePPIReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PPIResDto"
                }
              }
            }
          }
        },
        "tags": [
          "ppis"
        ]
      },
      "get": {
        "operationId": "PPIController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "classPeriod",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResDto"
                }
              }
            }
          }
        },
        "tags": [
          "ppis"
        ]
      }
    },
    "/ppis/{id}": {
      "get": {
        "operationId": "PPIController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PPIResDto"
                }
              }
            }
          }
        },
        "tags": [
          "ppis"
        ]
      },
      "put": {
        "operationId": "PPIController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePPIReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PPIResDto"
                }
              }
            }
          }
        },
        "tags": [
          "ppis"
        ]
      },
      "patch": {
        "operationId": "PPIController_updateSubjectPPIById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectPPIReqDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "ppis"
        ]
      },
      "delete": {
        "operationId": "PPIController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "ppis"
        ]
      }
    },
...
  },
  "info": {
    "title": "SIPPIE",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "MetadataDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "itemsPerPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "itemsPerPage",
          "totalPages",
          "totalItems"
        ]
      },
      "PaginationResDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataDto"
          }
        },
        "required": [
          "items",
          "metadata"
        ]
      },
...
      "CreateCourseReqDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "technologicalAxis": {
            "type": "string"
          },
          "educationLevel": {
            "type": "string",
            "example": "Superior"
          },
          "degree": {
            "type": "string",
            "example": "Bacharelado"
          },
          "modality": {
            "type": "string",
            "example": "Presencial"
          },
          "shift": {
            "type": "string",
            "example": "Noturno"
          }
        },
        "required": [
          "name",
          "technologicalAxis",
          "educationLevel",
          "degree",
          "modality",
          "shift"
        ]
      },
      "randomUUID": {
        "type": "object",
        "properties": {}
      },
      "CourseResDto": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/randomUUID"
          },
          "name": {
            "type": "string"
          },
          "technologicalAxis": {
            "type": "string"
          },
          "educationLevel": {
            "type": "string"
          },
          "degree": {
            "type": "string"
          },
          "modality": {
            "type": "string"
          },
          "shift": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "technologicalAxis",
          "educationLevel",
          "degree",
          "modality",
          "shift",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCourseReqDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "technologicalAxis": {
            "type": "string"
          },
          "educationLevel": {
            "type": "string"
          },
          "degree": {
            "type": "string"
          },
          "modality": {
            "type": "string"
          },
          "shift": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "technologicalAxis",
          "educationLevel",
          "degree",
          "modality",
          "shift"
        ]
      },
      "LoginReqDto": {
        "type": "object",
        "properties": {
          "registration": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "registration",
          "password"
        ]
      },
      "LoginResDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "expiresIn"
        ]
      },
      "SelectCourseReqDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string"
          }
        },
        "required": [
          "courseId"
        ]
      },
...
      "CreatePPIReqDto": {
        "type": "object",
        "properties": {
          "classPeriod": {
            "type": "string"
          },
          "workload": {
            "type": "number"
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "classPeriod",
          "workload",
          "subjects"
        ]
      },
      "PPIResDto": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/randomUUID"
          },
          "classPeriod": {
            "type": "string"
          },
          "workload": {
            "type": "number"
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "classPeriod",
          "workload",
          "subjects",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePPIReqDto": {
        "type": "object",
        "properties": {
          "classPeriod": {
            "type": "string"
          },
          "workload": {
            "type": "number"
          }
        },
        "required": [
          "classPeriod",
          "workload"
        ]
      },
      "UpdateSubjectPPIReqDto": {
        "type": "object",
        "properties": {
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "subjects"
        ]
      },
...
    }
  }
}